% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{bigcor}
\alias{bigcor}
\title{calculate pearson correlation using expression matrix, especially for big data.}
\usage{
bigcor(x, y = NULL, fun = c("cor", "cov"), size = 500,
  verbose = TRUE, ...)
}
\value{
correlation efficients matrix
}
\description{
calculate pearson correlation using expression matrix, especially for big data.
}
