% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{checkPoint}
\alias{checkPoint}
\title{To compare the consistency of `corTest` and `cor.test`}
\usage{
checkPoint(x, y = NULL, outputdf, rownum = NULL)
}
\description{
To compare the consistency of `corTest` and `cor.test`
}
