% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppServer.R
\docType{data}
\name{circbank}
\alias{circbank}
\title{shinyAppServer}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 140790 rows and 8 columns.}
\usage{
circbank
}
\description{
shinyAppServer
}
\keyword{datasets}
