% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corHeatmap.R
\name{corHeatmap}
\alias{corHeatmap}
\title{plot heatmap with normalized expression}
\usage{
corHeatmap(cor_table, circ_matrix, gene_matrix, posNum = 50,
  negNum = 50, needLog = TRUE)
}
\description{
plot heatmap with normalized expression
}
