% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{corMatrix2LongTable}
\alias{corMatrix2LongTable}
\title{To transform `pearsonTestMatrix` list to a long table}
\usage{
corMatrix2LongTable(lst, symmetric = TRUE)
}
\value{
a data.frame with correlation test results
}
\description{
To transform `pearsonTestMatrix` list to a long table
}
