% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goPathwayEnrichment.R
\name{goEn}
\alias{goEn}
\title{enrichment analysis for GO}
\usage{
goEn(gene, Ont = "BP", universeList = NULL, orgdb = NULL,
  keyType = "SYMBOL", minGeneNum = 5, maxGeneNum = 500,
  pvalueCutoff = 1, qvalueCutoff = 1, resultPvalueCutoff = 0.05,
  minOverlapNum = 3)
}
\description{
enrichment analysis for GO
}
