% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goPathwayEnrichment.R
\name{goPathwayEnrichment}
\alias{goPathwayEnrichment}
\title{the main function}
\usage{
goPathwayEnrichment(degList, universeList = NULL, species = "hsa",
  keyType = "SYMBOL", species_db = NULL, minGeneNum = 5,
  maxGeneNum = 500, pvalueCutoff = 1, qvalueCutoff = 1,
  resultPvalueCutoff = 0.05, minOverlapNum = 3)
}
\description{
the main function
}
