% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goPathwayEnrichment.R
\name{keggPath}
\alias{keggPath}
\title{enrichment analysis for KEGG}
\usage{
keggPath(gene, universeList = NULL, species = "hsa",
  keyType = "SYMBOL", minGeneNum = 5, maxGeneNum = 500,
  pvalueCutoff = 1, qvalueCutoff = 1, resultPvalueCutoff = 0.05,
  minOverlapNum = 3, use_internal_data = FALSE)
}
\value{
NULL

if (species == "hsa" && use_internal_data) {
  #' @note `quitely` must been add
  require(KEGG.db, quietly = TRUE)
  use_internal_data = TRUE
}
}
\description{
enrichment analysis for KEGG
}
