% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{pearsonCorrelationMatrix}
\alias{pearsonCorrelationMatrix}
\title{To recognize using `cor` or `bigcor`}
\usage{
pearsonCorrelationMatrix(x, y = NULL, size = 500)
}
\description{
To recognize using `cor` or `bigcor`
}
