% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{pearsonTestMatrix}
\alias{pearsonTestMatrix}
\title{A statistical test for correlation matrix with student t.test, please ref `cor.test.default`}
\usage{
pearsonTestMatrix(x, y = NULL, size = 500, adjust = "BH")
}
\value{
a list with coefficents, p values & fdr
}
\description{
A statistical test for correlation matrix with student t.test, please ref `cor.test.default`
}
