% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goPathwayEnrichment.R
\name{reacPath}
\alias{reacPath}
\title{enrichment analysis for reactome}
\usage{
reacPath(gene, universeList = NULL, species = "human",
  keyType = "SYMBOL", minGeneNum = 5, maxGeneNum = 500,
  pvalueCutoff = 1, qvalueCutoff = 1, resultPvalueCutoff = 0.05,
  minOverlapNum = 3)
}
\description{
enrichment analysis for reactome
}
